requirements.txt
numpy==1.19.2
Pillow==8.2.0
scipy==1.5.2
